Begun reworking UIManager
This rework will draw the UI fixed to the camera at a closer distance to the camera. This will allow the camera to be zoomed in and zoom in further (up to +500.0f). Cooldowns will be moved to and handled by PlayerManager.cpp. UIManager will only query PlayerManager on the player state, not handle input.